+++ tests/Themia.Messaging.AspNetCore.Tests/RoundTripTests.cs
@@ -1,3 +1,4 @@
+using Themia.TestSupport;
 using System.Net;
 
 using Microsoft.AspNetCore.Builder;